Small bug fixes
The code now works properly when DISP_HEIGHT is set to None. Previously it crashed when checking if DISP_HEIGHT is not given and needs to be calculated.
Moved check_params function before calc_DISP_HEIGHT to properly show the warning message associated with DISP_HEIGHT and CANOPY_HEIGHT parameters.
phi_c parameter is now properly calculated during stable stratifications in K&M2001 method.
Fixed a crash in K&M2001 method caused by ustar being zero, which led to kappa being zero and causing a division by zero error.
Added a zL filtering criteria to Kljun2015 and K&M2001, which was forgotten previously.
Added a progress printing for K&M2001 calculation because it takes a loooong time to calculate footprints for 10 years of data and it would be nice to know the progress.